+++ demo/demo.cpp
@@ -433,7 +433,7 @@ public:
 				} else {
 					mvaddstrattr(0, 0, "Render OFF");
 				}
-				mvaddstrattr((h >> 1) + 2, 0, string("This realm: \"") + this->name + "'s\" (birth " + set_to_str(this->birth) + ", survival " + set_to_str(this->survival) + "), SLE " + to_str(t - t_last_emit, 1) + ", autoemit (" + to_str(this->autoemit_interval, 1) + ") " + (autoemit ? "ON" : "OFF") + ", InConnsN " + to_string(this->efunguz->in_connections_num()));
+				mvaddstrattr((h >> 1) + 2, 0, string("This realm: \"") + this->name + "'s\" (birth " + set_to_str(this->birth) + ", survival " + set_to_str(this->survival) + "), SLE " + to_str(t - t_last_emit, 1) + ", autoemit (" + to_str(this->autoemit_interval, 1) + ") " + (autoemit ? "ON" : "OFF") + ", InAbsorbing " + to_string(this->efunguz->in_absorbing_num()) + ", InPermitted " + to_string(this->efunguz->in_permitted_num()) + ", InAttempted " + to_string(this->efunguz->in_attempted_num()));
 				mvaddstrattr((h >> 1) + 3, 0, "Other realms: ");
 				for (int i_other = 0; i_other < this->others.size(); i_other++) {
 					const auto& that = this->others[i_other];